+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Fixes a bug affecting GPU transforms of Clenshaw and HealPix grids [#921](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/921)
+- Updated docs CairoMakie compatibility to "0.12, 0.13, 0.14, 0.15"
+- OneBandShortwave radiation implemented with diagnostic clouds [#888](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/888)
 - AMDGPU extension [#901](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/901)
 - Don't adapt rings in RingGrids on GPU [#920](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/920)
 - README updates [#915](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/915#pullrequestreview-3469109183)
@@ -23,6 +25,7 @@
 - Seasonal ocean and land climatology models GPU-ready [#892](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/892)
 - Zenith constructors [#889](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/889)
 - More general forcing of pressure, vorticity and divergence [#855](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/855)
+
 
 ## v0.17.3